+++ src/Network/Network/AzureFirewallPolicy/SetAzureFirewallPolicyCommand.cs
@@ -162,6 +162,12 @@ namespace Microsoft.Azure.Commands.Network
             HelpMessage = "The Private IP Range")]
         public string[] PrivateRange { get; set; }
 
+       [Parameter(
+            Mandatory = false,
+            HelpMessage = "Explicit Proxy Settings in Firewall Policy.")]
+        public PSAzureFirewallPolicyExplicitProxy ExplicitProxy { get; set; }
+
+
         private void AddPremiumProperties(PSAzureFirewallPolicy firewallPolicy)
         {
             firewallPolicy.Sku = new PSAzureFirewallPolicySku
@@ -245,6 +251,7 @@ namespace Microsoft.Azure.Commands.Network
                 this.UserAssignedIdentityId = this.IsParameterBound(c => c.UserAssignedIdentityId) ? UserAssignedIdentityId : (InputObject.Identity?.UserAssignedIdentities != null ? InputObject.Identity.UserAssignedIdentities?.First().Key : null);
                 this.SkuTier = this.IsParameterBound(c => c.SkuTier) ? SkuTier : (InputObject.Sku?.Tier != null ? InputObject.Sku.Tier : null);
                 this.PrivateRange = this.IsParameterBound(c => c.PrivateRange) ? PrivateRange : InputObject.PrivateRange;
+                this.ExplicitProxy = this.IsParameterBound(c => c.ExplicitProxy) ? ExplicitProxy : InputObject.ExplicitProxy;
 
                 var firewallPolicy = new PSAzureFirewallPolicy()
                 {
@@ -256,7 +263,8 @@ namespace Microsoft.Azure.Commands.Network
                     BasePolicy = this.BasePolicy != null ? new Microsoft.Azure.Management.Network.Models.SubResource(this.BasePolicy) : null,
                     DnsSettings = this.DnsSetting,
                     SqlSetting = this.SqlSetting,
-                    PrivateRange = this.PrivateRange
+                    PrivateRange = this.PrivateRange,
+                    ExplicitProxy = this.ExplicitProxy
                 };
 
                 AddPremiumProperties(firewallPolicy);
@@ -280,7 +288,8 @@ namespace Microsoft.Azure.Commands.Network
                     BasePolicy = BasePolicy != null ? new Microsoft.Azure.Management.Network.Models.SubResource(BasePolicy) : null,
                     DnsSettings = this.DnsSetting,
                     SqlSetting = this.SqlSetting,
-                    PrivateRange = this.PrivateRange
+                    PrivateRange = this.PrivateRange,
+                    ExplicitProxy = this.ExplicitProxy
                 };
 
                 AddPremiumProperties(firewallPolicy);